requirements.txt
beautifulsoup4>=4.11.1 
bs4>=0.0.1 
soupsieve>=2.3.2.post1
requests>=2.27.1 

urllib3>=1.26.9 
idna>=3.3 
charset_normalizer>=2.0.12 
cerifi>=2022-10-8

mysql-connector>=2.2.9
lxml>=4.8.0

PySocks>=1.7.1
async-generator>=1.10
attrs>=21.4.0
cffi>=1.15.0
cryptography>=37.0.2
h11>=0.13.0
outcome>=1.1.0
pyOpenSSL>=22.0.0
pycparser>=2.21
selenium>=4.1.5
sniffio>=1.2.0
sortedcontainers>=2.4.0
trio>=0.20.0
trio-websocket>=0.9.2
wsproto>=1.1.0